+++ common/controllers/lifecycle/manager.go
@@ -4,9 +4,9 @@
 package lifecycle
 
 import (
-	compspi "github.com/verrazzano/verrazzano-modules/common/helm_component/spi"
 	"github.com/verrazzano/verrazzano-modules/common/controllers/base/basecontroller"
 	spi "github.com/verrazzano/verrazzano-modules/common/controllers/base/spi"
+	compspi "github.com/verrazzano/verrazzano-modules/common/helm_component/action_spi"
 	vzplatform "github.com/verrazzano/verrazzano-modules/module-operator/apis/platform/v1alpha1"
 	ctrlruntime "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
