+++ common/helm_component/handlers/factory/factory.go
@@ -4,9 +4,9 @@
 package factory
 
 import (
+	compspi "github.com/verrazzano/verrazzano-modules/common/helm_component/action_spi"
 	"github.com/verrazzano/verrazzano-modules/common/helm_component/handlers/install"
 	"github.com/verrazzano/verrazzano-modules/common/helm_component/handlers/uninstall"
-	compspi "github.com/verrazzano/verrazzano-modules/common/helm_component/spi"
 )
 
 func NewLifeCycleComponent() compspi.LifecycleComponent {
